% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggexpression_heatmap.R
\name{gg_expression_heatmap}
\alias{gg_expression_heatmap}
\title{Expression Heatmap}
\usage{
gg_expression_heatmap(
  data_main,
  col_genes,
  col_samples,
  col_expression,
  genes_to_include = NULL
)
}
\arguments{
\item{genes_to_include}{}
}
\description{
Expression Heatmap
}
