% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalised_sample_level_plotting.R
\name{gg_sample_annotation}
\alias{gg_sample_annotation}
\title{gg_sample_annotation}
\usage{
gg_sample_annotation(
  data,
  col_sample,
  col_sort,
  interactive = TRUE,
  svg_width = 10,
  svg_heigt = 4,
  numerical_plot_height_ratio = 2
)
}
\arguments{
\item{col_sample}{}
}
\description{
gg_sample_annotation
}
\examples{
brca_metadata_csv <- system.file(package='oncoplotgg', "testdata/BRCA_tcgamutations_mc3_clinicaldata.csv")
brca_metadata_df <- read.csv(file = brca_metadata_csv, header=TRUE)
gg_sample_annotation(brca_metadata_df, 'Tumor_Sample_Barcode')

# Try experimenting with sorting - you can sort on multiple columns
brca_metadata_csv <- system.file(package='oncoplotgg', "testdata/BRCA_tcgamutations_mc3_clinicaldata.csv")
brca_metadata_df <- read.csv(file = brca_metadata_csv, header=TRUE)
gg_sample_annotation(brca_metadata_df, 'Tumor_Sample_Barcode', col_sort = c("vital_status", "age_at_initial_pathologic_diagnosis"))
}
