% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggoncoplot.R
\name{ggoncoplot}
\alias{ggoncoplot}
\title{GG oncoplot}
\usage{
ggoncoplot(
  data_main,
  col_genes,
  col_samples,
  col_mutation_type = NULL,
  col_tooltip = col_samples,
  topn = 10,
  show_sample_ids = FALSE,
  interactive = TRUE,
  interactive_svg_width = 10,
  interactive_svg_height = 6,
  genes_to_include = NULL,
  xlab_title = "Sample",
  ylab_title = "Gene",
  sample_annotation_df = NULL
)
}
\arguments{
\item{data_main}{data for main oncoplot (data.frame)}

\item{col_genes}{name of \strong{data_main} column containing gene names/symbols (string)}

\item{col_samples}{name of \strong{data_main} column containing sample identifiers (string)}

\item{col_mutation_type}{name of \strong{data_main} column describing mutation types (string)}

\item{col_tooltip}{name of \strong{data_main} column containing whatever information you want to display in (string)}

\item{topn}{how many of the top genes to visualise. Ignored if \code{genes_to_include} is supplied (number)}

\item{show_sample_ids}{show sample_ids_on_x_axis (flag)}
}
\value{
ggplot or ggiraph object if \code{interactive=TRUE}
}
\description{
GG oncoplot
}
\examples{
GBM
gbm_csv <- system.file(package='oncoplotgg', "testdata/GBM_tcgamutations_mc3_maf.csv")
gbm_df <- read.csv(file = gbm_csv, header=TRUE)
ggoncoplot(gbm_df, 'Hugo_Symbol', 'Tumor_Sample_Barcode', col_mutation_type = 'Variant_Classification')

# BRCA
brca_csv <- system.file(package='oncoplotgg', "testdata/BRCA_tcgamutations_mc3_maf.csv")
brca_df <- read.csv(file = brca_csv, header=TRUE)
ggoncoplot(brca_df, 'Hugo_Symbol', 'Tumor_Sample_Barcode', col_mutation_type = 'Variant_Classification')
}
