% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalised_sample_level_plotting.R
\name{heirarchical_ranking}
\alias{heirarchical_ranking}
\title{Ranking based on multiple data columns}
\usage{
heirarchical_ranking(data_ls)
}
\arguments{
\item{data_ls}{}
}
\description{
Ranking based on multiple data columns
}
